separated testing function from schedule and fixed check_answer function
